cloudBackend/CloudFunction/MachineLearningModule/getTourPackage/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-aiplatform